backend/requirements.txt
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
python-dotenv==1.1.0